Refactored the following plugins to meet the latest changes and features introduced in MySQL Shell 8.0.22:
  - schema
  - user
  - support

Also updated the README file and removed unnecessary code.
